Use Warehouse relation and BigDecimal rev
Change summary:
- DashboardStoreRankingDTO: change revenue type from Long to BigDecimal.
- InventoryHistory entity: make warehouse JoinColumn insertable and remove the redundant warehouseId field.
- InventoryHistoryServiceImpl: switch to using the Warehouse association (null-safe id mapping, lookup by warehouse id to set Warehouse on create) instead of direct warehouseId access.
- InventoryServiceImpl: set Warehouse object on history records instead of warehouseId.
- application.properties: update SendGrid API key entry (format/value changed).
- InventoryItemModal.jsx: remove unused icon imports and replace DollarSign icon with a VND label.
- Add multiple image assets under uploads/.

These changes consolidate warehouse handling via the entity relationship, improve revenue precision, and include small UI/config updates.

+++ RetailChainService/src/main/java/com/sba301/retailmanagement/dto/response/DashboardStoreRankingDTO.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -22,7 +24,7 @@ public class DashboardStoreRankingDTO {
     /**
      * Tổng doanh thu của cửa hàng (tính từ inventory documents).
      */
-    private Long revenue;
+    private BigDecimal revenue;
 
 }
